+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Include key ID in "key not found" error message.
 - Don't reject remote actors without public keys.
 - Don't reject remote actors with unsupported key types.
+- Extract authority from request URI when processing `@authority` RFC-9421 component.
 
 ## [5.2.0] - 2026-04-24